Rename truffle-deployer to @truffle/deployer
