+++ packages/deployer/src/actions/new.js
@@ -2,7 +2,7 @@ module.exports = function(contract, args, deployer) {
   return async function() {
     deployer.logger.log("Creating new instance of " + contract.contract_name);
     // Evaluate any arguments if they're promises
-    const newArgs = await Promise.all(args);
+    await Promise.all(args);
     return contract.new.apply(contract, args);
   };
 };
