+++ packages/truffle-core/package.json
@@ -55,7 +55,7 @@
     "truffle-contract-sources": "^0.1.5",
     "@truffle/debug-utils": "^1.0.18",
     "@truffle/debugger": "^5.0.26",
-    "truffle-deployer": "^3.0.32",
+    "@truffle/deployer": "^3.0.32",
     "truffle-init": "^1.0.7",
     "@truffle/provider": "^0.1.16",
     "@truffle/provisioner": "^0.1.5",
