+++ packages/truffle-migrate/migration.js
@@ -1,5 +1,5 @@
 const path = require("path");
-const Deployer = require("truffle-deployer");
+const Deployer = require("@truffle/deployer");
 const Require = require("@truffle/require");
 const Emittery = require("emittery");
 const { Web3Shim } = require("truffle-interface-adapter");
