+++ packages/truffle-migrate/package.json
@@ -21,7 +21,7 @@
     "emittery": "^0.4.0",
     "node-dir": "0.1.17",
     "truffle-config": "^1.1.20",
-    "truffle-deployer": "^3.0.32",
+    "@truffle/deployer": "^3.0.32",
     "truffle-interface-adapter": "^0.2.5",
     "web3": "1.2.1"
   },
